weather-producer/requirements.txt
requests==2.31.0
pydantic==1.10.12
pika==1.3.1
APScheduler==3.10.1
python-dotenv==1.0.0
